requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
pytz==2023.3.post1
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
